Consolidate NestedMultiSelect and MultiSelect into single component
